requirements.txt
telethon
pandas
python-dotenv
requests
matplotlib
seaborn
